Refactor code actions provider internals
